refactor: configをinternal内に配置
