+++ main.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	"github.com/Misoten-B/airship-backend/config"
 	_ "github.com/Misoten-B/airship-backend/docs"
+	"github.com/Misoten-B/airship-backend/internal/drivers/config"
 	"github.com/Misoten-B/airship-backend/internal/frameworks"
 	"github.com/Misoten-B/airship-backend/internal/frameworks/router"
 	v1 "github.com/Misoten-B/airship-backend/internal/frameworks/router/v1"
